Move complex composite views/notifiers to shvatka.views
Extract ComplexOrgNotifier, ComplexGameViewPreparer, ComplexGameLogWriter,
ComplexTeamNotifier and ComplexView from main_factory into a dedicated
shvatka/views module. main_factory keeps only the DI provider wiring and
imports the composites. Remove the web-push design note.
